Revert "Rename `server_tick` module into `tick`"
This reverts commit 691add171fc0ee17a18dc2e62c504a7bba7c1c86.
