+++ src/core/tick.rs
@@ -10,7 +10,7 @@ use serde::{Deserialize, Serialize};
 /// All operations on it are wrapping.
 ///
 /// See also [`ServerInitTick`](crate::client::ServerInitTick) and
-/// [`ServerTick`](crate::server::tick::ServerTick).
+/// [`ServerTick`](crate::server::server_tick::ServerTick).
 #[derive(Clone, Copy, Debug, Default, Deserialize, Eq, Hash, PartialEq, Serialize)]
 pub struct RepliconTick(u32);
