+++ tests/other.rs
@@ -1,6 +1,6 @@
 use bevy::prelude::*;
 use bevy_replicon::{
-    core::channels::ReplicationChannel, prelude::*, server::tick::ServerTick,
+    core::channels::ReplicationChannel, prelude::*, server::server_tick::ServerTick,
     test_app::ServerTestAppExt,
 };
 use serde::{Deserialize, Serialize};
